Fix for WFMP-235, Introduce a start-jar goal
